requirements.txt
fastapi=0.103.2
uvicorn=0.22.0
sqlalchemy=2.0.30
pyodbc=4.0.39
pandas=1.3.5
python-dotenv=0.21.1